+++ CJBCheatsMenu/release-notes.md
@@ -5,6 +5,7 @@
 * Added warp to the Skull Cavern (thanks to Enaium!).
 * Added button to reset controls to default.
 * Added support for disabling a control by pressing escape when 'Press New Key' is shown. (The open-menu key will reset to default instead, so you can't accidentally lock yourself out of the menu.)
+* Added option to increase grow crop/tree area around the player.
 * The 'freeze time' control is now unbound by default. (Players with an existing `config.json` aren't affected unless they reset controls to default.)
 * Fixed grow trees not working consistently after 1.22.
 * Fixed fast egg incubator not working.